// EcoDOC/2.Source/1.Development/2.Frontend/src/app/core/model/domain/Rules.ts
export class Rule {
    ticKet: string;
    nodeId: string;
    title: string;
    disabled: boolean;
    applyToChildren: boolean;
    executeAsynchronously: boolean;
    description: string;
    actionDefinitionName: string;
    conditionDefinitionName: string;
    actionsDefinitionName: string;
    typeName: string;
    ruleType: string;
    actionDefinitionNameCha: string

}